+++ Cosette/Engine/Ai/Score/EvaluationConstants.cs
@@ -13,7 +13,12 @@
         public static int[] IsolatedPawns = { -20, -10 };
         public static int[] ChainedPawns = { 5, 5 };
         public static int[] PassingPawns = { 10, 50 };
-        public static int[] Mobility = { 4, 1 };
+
+        public static int[] Mobility = { 3, 1 };
+        public static int CenterMobilityModifier = 3;
+        public static int ExtendedCenterMobilityModifier = 2;
+        public static int OutsideMobilityModifier = 1;
+
         public static int[] KingInDanger = { -20, -5 };
         public static int[] PawnShield = { 20, 0 };
         public static int[] DoubledRooks = { 40, 10 };
@@ -21,5 +26,9 @@
         public static int[] PairOfBishops = { 30, 20 };
 
         public const int OpeningEndgameEdge = 20500;
+
+        public const ulong Center = 0x1818000000;
+        public const ulong ExtendedCenter = 0x3c24243c0000;
+        public const ulong Outside = 0xffffc3c3c3c3ffff;
     }
 }